LayoutTests/fast/js/inc-const-valueOf-expected.txt

Test for regression against

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testPostIncConstVarWithIgnoredResult() is true
PASS testPreIncConstVarWithIgnoredResult() is true
PASS testPreIncConstVarWithAssign() is true
PASS successfullyParsed is true

TEST COMPLETE
